requirements.txt
fastf1
scikit-learn
pandas
matplotlib
python-dotenv
requests